Uploading files of progress so far
I created these files without the use of NetBeans because I didn't want to use an unknown software that may end up doing something wrong or strange. The server and servergui work together and the client and clientgui work together. The message, messagelist and node work together to format/parse and save a list of messages sent and received. I have not programmed the game yet, just getting the server and client connected and determining if they are ready to play.